requirements.txt
google-auth
google-auth-oauthlib
google-api-python-client
requests
python-dotenv
moviepy==1.0.3
numpy
soundfile
ollama
kokoro
torch
whisperx